requirements.txt
flask==2.3.3
redis==5.0.1
python-dotenv==1.0.0
flask-swagger-ui==4.11.1
requests==2.31.0
gunicorn==21.2.0
werkzeug==2.3.7